move project types
